fixes #142 alternatives with whitespace in overlong names
in the UI we prevent alternative names longer then 30 characters
But if alternatives are created from services, they might be too long
and have to be truncated.
If the name contains whitespaces, it could happen that the truncated name had trailing whitespace,
this can cause troubles ( see #169 ).
Now truncated names are trimmed.